Update README.md
In this assignment, you will implement learning and inference procedures for some of the proba-bilistic models described in class, apply your solutions to some simulated datasets, and analyze the results.

General Note:

Full points are given for complete solutions, including justifying the choices or assumptions you made to solve each question.

A written report should be included in the nal submission. Do not dump your codes and outputs in the report. Keep it short, readable, and well-organized.

Programming assignments are to be solved and submitted individually. You are encouraged to discuss the assignment with other students, but you must solve it on your own.

Please ask all questions related to this assignment on Piazza, using the tag pa3.

There are 3 starter les attached, helper.py, starter_kmeans.py and starter_gmm.py which will help you with your implementation.

1 K-MEANS [9 PT.]

K-means [9 pt.]